+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - 0.1.3
     - Optimized Cosine distance kernel
+    - Optimized (NaN) Safe Euclidean distance kernel
 
 - 0.1.2
     - Fixed Grid Search best hyper-parameters method